feat(explorer): rebrand epoch page: checkpoints. Move tabs to ui-kit.
