+++ apps/ui-kit/src/lib/components/atoms/tabs/index.ts
@@ -1,0 +1,7 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './Tabs';
+export * from './TabsContent';
+export * from './TabsHeader';
+export * from './TabsTrigger';
